Tune.Link (tune v0.1.0)

Provides functions to generate integration links from tracks, artists and albums.

Implementations are extremely naive and rely on building URLs from metadata, but there are times when they simply don't work, particularly with titles that have suffixes like "2017 remaster" or "feat. another artist name".

Functions

last_fm(album)

Specs

last_fm(Tune.Spotify.Schema.Album.t() | Tune.Spotify.Schema.Artist.t()) ::
  String.t()

last_fm(track, album, artist)

Specs

last_fm(
  Tune.Spotify.Schema.Track.t(),
  Tune.Spotify.Schema.Album.t(),
  Tune.Spotify.Schema.Artist.t()
) :: String.t()

musixmatch(track, artist)

Specs

musixmatch(Tune.Spotify.Schema.Track.t(), Tune.Spotify.Schema.Artist.t()) ::
  String.t()

wikipedia(artist)

Specs

wikipedia(Tune.Spotify.Schema.Artist.t()) :: String.t()

youtube(album)

Specs

youtube(Tune.Spotify.Schema.Album.t() | Tune.Spotify.Schema.Artist.t()) ::
  String.t()

youtube(track, artist)

Specs

youtube(Tune.Spotify.Schema.Track.t(), Tune.Spotify.Schema.Artist.t()) ::
  String.t()