tune v0.1.0 TuneWeb.AuthView

Functions

__resource__()

The resource name, as an atom, for this view

render(template, assigns \\ %{})

Renders the given template locally.

template_not_found(template, assigns)

Specs

template_not_found(Phoenix.Template.name(), map()) :: no_return()

Callback invoked when no template is found. By default it raises but can be customized to render a particular template.