TuneWeb.Router.Helpers (tune v0.1.0)

Module with named helpers generated from TuneWeb.Router.

Functions

auth_path(conn_or_endpoint, action)

auth_path(conn_or_endpoint, action, params)

auth_path(conn_or_endpoint, action, provider, params)

auth_url(conn_or_endpoint, action)

auth_url(conn_or_endpoint, action, params)

auth_url(conn_or_endpoint, action, provider, params)

explorer_path(conn_or_endpoint, action)

explorer_path(conn_or_endpoint, action, params)

explorer_path(conn_or_endpoint, action, artist_id, params)

explorer_url(conn_or_endpoint, action)

explorer_url(conn_or_endpoint, action, params)

explorer_url(conn_or_endpoint, action, artist_id, params)

live_dashboard_path(conn_or_endpoint, action)

live_dashboard_path(conn_or_endpoint, action, params)

live_dashboard_path(conn_or_endpoint, action, node, page)

live_dashboard_path(conn_or_endpoint, action, node, page, params)

live_dashboard_url(conn_or_endpoint, action)

live_dashboard_url(conn_or_endpoint, action, params)

live_dashboard_url(conn_or_endpoint, action, node, page)

live_dashboard_url(conn_or_endpoint, action, node, page, params)

path(data, path)

Generates the path information including any necessary prefix.

static_integrity(endpoint, path)

Generates an integrity hash to a static asset given its file path.

static_path(conn, path)

Generates path to a static asset given its file path.

static_url(conn, path)

Generates url to a static asset given its file path.

url(data)

Generates the connection/endpoint base URL without any path information.